+++ src/components/test/InstructionAudio.tsx
@@ -1,0 +1,14 @@
+import * as React from "react";
+import Button from "~/components/Button";
+
+function InstructionAudio() {
+  return (
+    <>
+      <h1>Press the following button and listen to the sentence.</h1>
+
+      <Button>Listen</Button>
+    </>
+  );
+}
+
+export default InstructionAudio;
